+++ app/Http/Controllers/API/V1/Controller.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\API\V1;
 
 abstract class Controller
 {
