+++ app/Http/Controllers/API/V1/LatestPosts.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\API\V1;
 
 use App\Http\Resources\PostResource;
 use App\Models\Post;
